Consolidate legacy checkpoint scripts into folder
