+++ docs/guides/knowledge_distillation.md
@@ -71,7 +71,7 @@ MaxText requires checkpoints to be in a specific format. You'll need to convert 
 ```bash
 # Get unscanned checkpoint for efficient decoding
 JAX_PLATFORMS=cpu \
-python3 -m MaxText.convert_deepseek_family_unscanned_ckpt \
+python3 -m MaxText.utils.ckpt_scripts.convert_deepseek_family_unscanned_ckpt \
   --base_model_path ~/deepseek2-16b-chat \
   --maxtext_model_path ${BASE_DIRECTORY}/deepseek2-16-chat/unscanned \
   --model_size deepseek2-16b
@@ -91,7 +91,7 @@ MaxText requires checkpoints to be in a specific format. You'll need to convert 
 ```bash
 # Get scanned checkpoint for fine-tuning
 JAX_PLATFORMS=cpu \
-python3 -m MaxText.llama_or_mistral_ckpt \
+python3 -m MaxText.utils.ckpt_scripts.llama_or_mistral_ckpt \
   --base-model-path ~/llama2-7b-chat \
   --maxtext-model-path ${BASE_DIRECTORY}/llama2-7b-chat/scanned \
   --model-size llama2-7b
