+++ end_to_end/tpu/deepseek/v3-671b/test_deepseek_mtp.sh
@@ -16,7 +16,7 @@ export TOKENIZER_PATH='deepseek-ai/DeepSeek-V3'
 # The --enable_mtp flag is crucial for processing the MTP-specific layers.
 export CHKPT_BUCKET=gs://maxtext-deepseek/deepseek3-671b/hf
 export MODEL_BUCKET=gs://maxtext-deepseek/deepseek3-671b
-JAX_PLATFORMS=cpu python3 -m MaxText.convert_deepseek_family_ckpt \
+JAX_PLATFORMS=cpu python3 -m MaxText.utils.ckpt_scripts.convert_deepseek_family_ckpt \
     --base_model_path ${CHKPT_BUCKET} \
     --maxtext_model_path ${MODEL_BUCKET}/${idx} \
     --model_size ${MODEL_NAME} \
