+++ end_to_end/tpu/test_gpt3.sh
@@ -6,7 +6,7 @@ export OUTPUT_PATH=gs://maxtext-gpt3/tests
 export RUN_NAME=test_${TIMESTAMP}
 
 # convert gpt3-52k model
-python3 -m MaxText.convert_gpt3_ckpt_from_paxml --paxml-ckpt-path=${PAXML_CKPT_PATH} --maxtext-model-name=gpt3-52k --run-name=${RUN_NAME} --base-output-directory=${OUTPUT_PATH}
+python3 -m MaxText.utils.ckpt_scripts.convert_gpt3_ckpt_from_paxml --paxml-ckpt-path=${PAXML_CKPT_PATH} --maxtext-model-name=gpt3-52k --run-name=${RUN_NAME} --base-output-directory=${OUTPUT_PATH}
 
 # Run gpt3-52k with the converted ckpt
 python3 -m MaxText.train "${MAXTEXT_PKG_DIR:-${MAXTEXT_REPO_ROOT:-$PWD}/src/MaxText}"/configs/base.yml run_name=${RUN_NAME} model_name=gpt3-52k\
