+++ src/MaxText/utils/ckpt_scripts/convert_deepseek_family_ckpt.py
@@ -18,7 +18,7 @@ This script now includes support for converting Multi-Token Prediction (MTP) wei
 
 Example cmd:
 
-python3 -m MaxText.convert_deepseek_family_ckpt --base_model_path <path/to/meta/ckpt> \
+python3 -m MaxText.utils.ckpt_scripts.convert_deepseek_family_ckpt --base_model_path <path/to/meta/ckpt> \
     --maxtext_model_path <GCS/path/to/save/new/maxtext/ckpt> --model_size deepseek3-671b --enable_mtp
 """
