+++ src/MaxText/utils/ckpt_scripts/convert_deepseek_family_unscanned_ckpt.py
@@ -16,7 +16,7 @@ r"""Convert weights from a DeepSeek style model to a MaxText one.
 
 Example cmd:
 
-python3 -m MaxText.convert_deepseek_family_unscanned_ckpt --base_model_path <path/to/meta/ckpt> \
+python3 -m MaxText.utils.ckpt_scripts.convert_deepseek_family_unscanned_ckpt --base_model_path <path/to/meta/ckpt> \
     --maxtext_model_path <GCS/path/to/save/new/maxtext/ckpt> --model_size deepseek2-16b
 """
 
@@ -35,7 +35,7 @@ import torch
 import psutil
 from tqdm import tqdm
 
-from MaxText import convert_deepseek_family_ckpt as ds_ckpt
+from MaxText.utils.ckpt_scripts import convert_deepseek_family_ckpt as ds_ckpt
 from MaxText import llama_or_mistral_ckpt
 from MaxText import max_logging
 from MaxText.inference_utils import str2bool
