+++ src/MaxText/utils/ckpt_scripts/dequantize_mxfp4.py
@@ -16,8 +16,8 @@
 
 Example cmd:
 
-python3 dequantize_mxfp4.py --input-path=<input_path> --output-path=<output_path>
-python3 dequantize_mxfp4.py --input-path=<input_path> --output-path=<output_path> --dtype-str=bf16 --cache-size=2
+python3 -m MaxText.utils.ckpt_scripts.dequantize_mxfp4 --input-path=<input_path> --output-path=<output_path>
+python3 -m MaxText.utils.ckpt_scripts.dequantize_mxfp4 --input-path=<input_path> --output-path=<output_path> --dtype-str=bf16 --cache-size=2
 """
 
 import os
