+++ src/MaxText/utils/ckpt_scripts/llama_ckpt_conversion_inference_only.py
@@ -20,7 +20,7 @@ Get LLaMA pytorch_vars from Meta
 
 Example cmd:
 To save a ckpt
-python3 -m MaxText.llama_or_mistral_ckpt --base-model-path <path/to/meta/ckpt> \
+python3 -m MaxText.utils.ckpt_scripts.llama_or_mistral_ckpt --base-model-path <path/to/meta/ckpt> \
     --maxtext-model-path <GCS/path/to/save/new/maxtext/ckpt> --model-size llama2-7b
 
 For large size model (e.g. 70B model), this script requires large memory VM.
